Move filtering code to 'filter' directory
