+++ src/lib/filter/filter.test.ts
@@ -4,9 +4,9 @@
 import { expect, describe, it } from "@jest/globals";
 import { kind, modelToGroupVersionKind } from "kubernetes-fluent-client";
 import * as fc from "fast-check";
-import { CreatePod, DeletePod } from "../fixtures/loader";
+import { CreatePod, DeletePod } from "../../fixtures/loader";
 import { shouldSkipRequest } from "./filter";
-import { AdmissionRequest, Binding, Event } from "./types";
+import { AdmissionRequest, Binding, Event } from "../types";
 
 export const callback = () => undefined;
